Put dependency-free websockets protocol in websockets_protocol.py
